+++ Editor/DeviceDatabase.cs
@@ -1,13 +1,27 @@
 using System.IO;
 using Boo.Lang;
-using UnityEditor;
 using UnityEngine;
 
 namespace Unity.DeviceSimulator
 {
+    internal class DeviceHandle
+    {
+        public DeviceHandle(int id, string name, string directory)
+        {
+            Id = id;
+            Name = name;
+            Directory = directory;
+        }
+
+        public int Id { get; }
+        public string Name { get; }
+        public string Directory { get; }
+    }
+
     internal class DeviceDatabase
     {
         private List<DeviceInfo> m_Devices = new List<DeviceInfo>();
+        private readonly List<DeviceHandle> m_DeviceHandles = new List<DeviceHandle>();
 
         public DeviceDatabase()
         {
@@ -17,28 +31,52 @@ namespace Unity.DeviceSimulator
         public void Refresh()
         {
             m_Devices.Clear();
+            m_DeviceHandles.Clear();
 
-            var deviceInfoFilePath = Path.Combine("Packages", "com.unity.device-simulator", "DeviceDefinitions");
-            foreach (var deviceInfoFiles in Directory.GetFiles(deviceInfoFilePath, "*.json", SearchOption.TopDirectoryOnly))
+            var deviceDirectoryPaths = new[]
             {
-                var deviceInfo = JsonUtility.FromJson<DeviceInfo>(File.ReadAllText(deviceInfoFiles));
-                // TODO could improve in the future that only loads the texture while it's the current device.
-                if (!string.IsNullOrEmpty(deviceInfo.Meta.overlay))
-                {
-                    deviceInfo.Meta.overlayImage = AssetDatabase.LoadAssetAtPath<Texture>(Path.Combine(deviceInfoFilePath, deviceInfo.Meta.overlay));
-                }
-                else
-                {
-                    deviceInfo.Meta.overlayOffset = new Vector4(40, 60, 40, 60);
-                }
+                Path.Combine("Packages", "com.unity.device-simulator", ".DeviceDefinitions"),
+                DeviceSimulatorUserSettingsProvider.LoadOrCreateSettings().DeviceDirectory
+            };
 
-                m_Devices.Add(deviceInfo);
+            foreach (var deviceDirectoryPath in deviceDirectoryPaths)
+            {
+                if (string.IsNullOrEmpty(deviceDirectoryPath) || !Directory.Exists(deviceDirectoryPath))
+                    continue;
+
+                var deviceDirectory = new DirectoryInfo(deviceDirectoryPath);
+                var deviceDefinitions = deviceDirectory.GetFiles("*.device.json");
+
+                foreach (var deviceDefinition in deviceDefinitions)
+                {
+                    DeviceInfo deviceInfo;
+                    using (StreamReader sr = deviceDefinition.OpenText())
+                    {
+                        deviceInfo = JsonUtility.FromJson<DeviceInfo>(sr.ReadToEnd());
+                    }
+
+                    m_Devices.Add(deviceInfo);
+                    m_DeviceHandles.Add(new DeviceHandle(m_Devices.Count - 1, deviceInfo.Meta.friendlyName, deviceDirectory.FullName));
+                }
             }
+            m_DeviceHandles.Sort((x, y) => string.CompareOrdinal(x.Name, y.Name));
         }
 
-        public DeviceInfo[] GetDevices()
+        public DeviceInfo GetDevice(DeviceHandle handle)
         {
-            return m_Devices.ToArray();
+            var deviceInfo =  m_Devices[handle.Id];
+
+            if (!deviceInfo.LoadOverlayImage(handle) && deviceInfo.Meta.overlayOffset == Vector4.zero)
+            {
+                deviceInfo.Meta.overlayOffset = new Vector4(40, 60, 40, 60);
+            }
+
+            return deviceInfo;
+        }
+
+        public DeviceHandle[] GetDeviceHandles()
+        {
+            return m_DeviceHandles.ToArray();
         }
     }
 }