+++ Editor/settings/DeviceSimulatorUserSettings.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Unity.DeviceSimulator
+{
+    internal class DeviceSimulatorUserSettings : ScriptableObject
+    {
+        [SerializeField] public string DeviceDirectory;
+    }
+}